Imported Upstream version 3.10.7

+++ tests/baseclass.py
@@ -27,13 +27,16 @@ import unittest
 # Base class for any test case that requires a temp device node
 class RequiresDeviceNode(unittest.TestCase):
     def setUp(self):
+        self.addCleanup(self.removeTempDevice)
+
         (fd, self.path) = tempfile.mkstemp(prefix="temp-device-")
         f = os.fdopen(fd)
         f.seek(140000)
         os.write(fd, b"0")
 
-    def tearDown(self):
-        os.unlink(self.path)
+    def removeTempDevice(self):
+        if self.path and os.path.exists(self.path):
+            os.unlink(self.path)
 
 # Base class for any test case that requires a _ped.Device or parted.Device
 # object first.
@@ -46,6 +49,8 @@ class RequiresDevice(RequiresDeviceNode):
 # Base class for any test case that requires a filesystem on a device.
 class RequiresFileSystem(unittest.TestCase):
     def setUp(self):
+        self.addCleanup(self.removeTempDevice)
+
         self._fileSystemType = {}
         ty = _ped.file_system_type_get_next()
         self._fileSystemType[ty.name] = ty
@@ -63,13 +68,14 @@ class RequiresFileSystem(unittest.TestCase):
         os.write(fd, b"0")
         f.close()
 
-        os.system("/sbin/mke2fs -F -q %s" % (self.path,))
+        os.system("mke2fs -F -q %s" % (self.path,))
 
         self._device = _ped.device_get(self.path)
         self._geometry = _ped.Geometry(self._device, 0, self._device.length - 1)
 
-    def tearDown(self):
-        os.unlink(self.path)
+    def removeTempDevice(self):
+        if self.path and os.path.exists(self.path):
+            os.unlink(self.path)
 
 # Base class for certain alignment tests that require a _ped.Device
 class RequiresDeviceAlignment(RequiresDevice):
@@ -127,7 +133,7 @@ class RequiresDeviceAlignment(RequiresDevice):
 class RequiresLabeledDevice(RequiresDevice):
     def setUp(self):
         RequiresDevice.setUp(self)
-        os.system("/sbin/parted -s %s mklabel msdos" % (self.path,))
+        os.system("parted -s %s mklabel msdos" % (self.path,))
 
 # Base class for any test case that requires a _ped.Disk or parted.Disk.
 class RequiresDisk(RequiresDevice):
@@ -139,20 +145,21 @@ class RequiresDisk(RequiresDevice):
 # Base class for any test case that requires a filesystem made and mounted.
 class RequiresMount(RequiresDevice):
     def setUp(self):
+        self.addCleanup(self.removeMountpoint)
         RequiresDevice.setUp(self)
         self.mountpoint = None
 
     def mkfs(self):
-        os.system("/sbin/mkfs.ext2 -F -q %s" % self.path)
+        os.system("mkfs.ext2 -F -q %s" % self.path)
 
     def doMount(self):
         self.mountpoint = tempfile.mkdtemp()
-        os.system("/sbin/mount -o loop %s %s" % (self.path, self.mountpoint))
+        os.system("mount -o loop %s %s" % (self.path, self.mountpoint))
 
-    def tearDown(self):
-        os.system("/sbin/umount %s" % self.mountpoint)
-        os.rmdir(self.mountpoint)
-        RequiresDevice.tearDown(self)
+    def removeMountpoint(self):
+        if self.mountpoint and os.path.exists(self.mountpoint):
+            os.system("umount %s" % self.mountpoint)
+            os.rmdir(self.mountpoint)
 
 # Base class for any test case that requires a _ped.Partition.
 class RequiresPartition(RequiresDisk):
